BACKPORT: FROMLIST: PM / OPP: Support adjusting OPP voltages at runtime
On some SoCs the Adaptive Voltage Scaling (AVS) technique is
employed to optimize the operating voltage of a device. At a
given frequency, the hardware monitors dynamic factors and either
makes a suggestion for how much to adjust a voltage for the
current frequency, or it automatically adjusts the voltage
without software intervention. Add an API to the OPP library for
the former case, so that AVS type devices can update the voltages
for an OPP when the hardware determines the voltage should
change. The assumption is that drivers like CPUfreq or devfreq
will register for the OPP notifiers and adjust the voltage
according to suggestions that AVS makes.

Cc: Nishanth Menon <nm@ti.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
(am from https://patchwork.kernel.org/patch/7221511/)
Signed-off-by: Pi-Cheng Chen <pi-cheng.chen@mediatek.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
[djkurtz: Changes to drivers/base/power/opp.c applied to
 drivers/base/power/opp/core.c due to an UPSTREAM
 refactoring after this ancient FROMLIST patch was last
 updated. ]

BUG=chrome-os-partner:50411
TEST=build and boot on oak-rev5

Change-Id: I017c4277051928c2b3972cc93cac82294f1a356a
Reviewed-on: https://chromium-review.googlesource.com/330381
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>