UPSTREAM: r8152: add reset_resume function
When the reset_resume() is called, the flag of SELECTIVE_SUSPEND should be
cleared and reinitialize the device, whether the SELECTIVE_SUSPEND is set
or not. If reset_resume() is called, it means the power supply is cut or the
device is reset. That is, the device wouldn't be in runtime suspend state and
the reinitialization is necessary.

Signed-off-by: Hayes Wang <hayeswang@realtek.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 7ec2541aa7a026874930d6117e86fe2e76fb1587)

BUG=chrome-os-partner:53802
TEST=built/installed on gale and sentry. iperf no longer TX timeouts.

Change-Id: I052f3a46965156c20be1bf1001e76459b5124d9b
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348932
Commit-Ready: Zhifeng Cai <caiz@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>