CHROMIUM: drm/i915: Allow a way to disable watermark for debuging purposes.
Without watermark the power consumption will blow up, but
when enabling platforms and dealing with different kinds of
crashes, screen corruptions, pipe underuns, etc we need to be
able to easily disable watermark to see if we are on the right
investigation track.

Another possibility was to skip at the beginning and avoid all
calculations, but I'm not sure about it. Maybe it might be still
useful to know the calculated values when debuging.

V2:
If i915.enable_watermark=2, enable watermark in single pipe config and
disable watermark in multi pipe config.

BUG=chrome-os-partner:45364,chrome-os-partner:48749,chrome-os-partner:50117
TEST=Check the wm registers running "intel_reg dump | grep WM"

(am from https://patchwork.freedesktop.org/patch/69417/)
Change-Id: Ib7af0408aa04cd2d68679882b7dda14e0b3de5f9
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/334930
Tested-by: Prathyushi Nangia <prathyushi.nangia@intel.com>
Reviewed-by: Todd Broch <tbroch@chromium.org>