UPSTREAM: ALSA: timer: Fix race among timer ioctls
ALSA timer ioctls have an open race and this may lead to a
use-after-free of timer instance object.  A simplistic fix is to make
each ioctl exclusive.  We have already tread_sem for controlling the
tread, and extend this as a global mutex to be applied to each ioctl.

The downside is, of course, the worse concurrency.  But these ioctls
aren't to be parallel accessible, in anyway, so it should be fine to
serialize there.

Reported-by: Dmitry Vyukov <dvyukov@google.com>
Tested-by: Dmitry Vyukov <dvyukov@google.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
(cherry picked from commit af368027a49a751d6ff4ee9e3f9961f35bb4fede)
Signed-off-by: Stephen Barber <smbarber@chromium.org>

BUG=b:28694392
TEST=kernel compiles

Change-Id: I25db36e012d7db63559041564029f9b6ed54fe7a
Reviewed-on: https://chromium-review.googlesource.com/353723
Commit-Ready: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Dmitry Grinberg <dmitrygr@google.com>