CHROMIUM: spi: qup: Fix transaction done signaling
Wait to signal done until we get all of the interrupts we are expecting
to get for a transaction.  If we don't wait for the input done flag, we
can be inbetween transactions when the done flag comes in and this can
mess up the next transaction.

CC: Grant Grundler <grundler@chromium.org>
CC: Sarthak Kukreti <skukreti@codeaurora.org>
Signed-off-by: Andy Gross <andy.gross@linaro.org>
(cherry picked from commit 89fa07a49de4c9428a2ec272af87aa3ae45ab119
 git://git.kernel.org/pub/scm/linux/kernel/git/agross/linux.git)

BUG=chrome-os-partner:49255
TEST=SPI nor works in block mode

Change-Id: I903c41e0c40e00fd506132efd00ee95cf085ba07
Signed-off-by: Matthew McClintock <mmcclint@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/335625
Commit-Ready: Matthew McClintock <mmcclint@qca.qualcomm.com>
Tested-by: Matthew McClintock <mmcclint@qca.qualcomm.com>
Reviewed-by: Grant Grundler <grundler@chromium.org>