CHROMIUM: iwl7000: mac80211: provide interface to driver to set VHT MU-MIMO data
Provide an interface to the lower level driver to set the VHT
MU-MIMO data. This is needed for example when there is an update
of the group data during low power state, where the management
frame will not be passed to the host at all.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: b8dd8c6005fdb4e02b455ef8691c913ccdbdf19b