ANDROID: vfs: add d_canonical_path for stacked filesystem support
Inotify does not currently know when a filesystem
is acting as a wrapper around another fs. This means
that inotify watchers will miss any modifications to
the base file, as well as any made in a separate
stacked fs that points to the same file.
d_canonical_path solves this problem by allowing the fs
to map a dentry to a path in the lower fs. Inotify
can use it to find the appropriate place to watch to
be informed of all changes to a file.

Signed-off-by: Daniel Rosenberg <drosen@google.com>
(cherry picked from commit 96fe5c2548e745d2b2a95dd5cf800d4e36f81d26)

BUG=b:27794037
TEST=Build and boot on Smaug.

Change-Id: I063ea9c687df919c8e56046584ca9ebcd68c2a00
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/335184
Reviewed-by: Daniel Rosenberg <drosen@google.com>