CHROMIUMOS: drm/img-rogue: Disable firmware option check
When doing firmware load, we do a check to ensure that the user libraries
and firmware are compatible.  But when SUPPORT_KERNEL_SRVINIT is enabled,
this is not very useful - instead of comparing the user options to the
firmware options, we are comparing the kernel options to the firmware
options, rendering the check wrong and meaningless.  (In particular, KM
options are allowed to be different from UM/FW options.)

So disable the check when SUPPORT_KERNEL_SRVINIT is enabled.  In a future
DDK release, we will fix this check so it is not done in the kernel.

BUG=chrome-os-partner:51192
TEST=Build debug UM and release KM.  Boot to UI.

Change-Id: I43e3d1f75960d9eb0c6c2b74e5beec4e629b9f01
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/332601
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>