CHROMIUM: LSM: make modules FS pinning work with unshared mount namespace
Currently pinning modules filesystem is done by storing pointer to
"struct vfsmount" and comparing it to vfsmount the module is loaded from.
This does not work in presence of private mount namespaces.

This patch changes Chromium LSM so that instead of vfsmount it keeps a pointer
to super_block. Super block is unique for each mounted filesystem and stays
the same across private mount namespacess.

BUG=b:27720867
TEST=modules load from "unshared" mount namespace without errors

Change-Id: I41be0c990ce8af578d6ae39c833d0f2fbe98bcb6
Signed-off-by: Andrey Ulanov <andreyu@google.com>
Reviewed-on: https://chromium-review.googlesource.com/334911
Reviewed-by: Kees Cook <keescook@chromium.org>