CHROMIUM: sdhci: tegra: Reenable 64-bit ADMA
64-bit ADMA was broken on Smaug due to the handling of buffers that do
not meet the 8-byte alignment requirement in the sdhci driver.  For the
record, Tegra can do 4-byte alignment.

The key issue was that if the DMA transfer length was less than the
alignment requirement, an extra DMA descriptor of transfer length zero
would be added.  Tegra does not handle this case well.  This problem has
been fixed upstreamed and with that fix, we can reenable 64-bit ADMA on
T210

BUG=chrome-os-partner:42771
TEST=Boot on Smaug

Change-Id: I9d301ed5f89b989eb9c74870a0ec2ae755d70abc
Signed-off-by: Christopher Freeman <cfreeman@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/335594
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>