UPSTREAM: ASoC: hdmi-codec: Add hdmi-codec for external HDMI-encoders
The hdmi-codec is a platform device driver to be registered from
drivers of external HDMI encoders with I2S and/or spdif interface. The
driver in turn registers an ASoC codec for the HDMI encoder's audio
functionality.

The structures and definitions in the API header are mostly redundant
copies of similar structures in ASoC headers. This is on purpose to
avoid direct dependencies to ASoC structures in video side driver.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Acked-by: Arnaud Pouliquen <arnaud.pouliquen@st.com>
Acked-by: PC Liao <pc.liao@mediatek.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 09184118a8abae030539469848d475adcc0e5839)

BUG=chrome-os-partner:40529
TEST=build and boot oak-tot to UI

Change-Id: I17eb40148f829b5f3648aa7ea682061024b3e162
Signed-off-by: PC Liao <pc.liao@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/320661
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>