CHROMIUM: drm/img-rogue: Honor cache invalidate calls
Due to speculative data prefetching done by some processors we need
to issue an invalidate (after device operation and) before accessing
the buffer touched by the device. Therefore, for processors that
support cache invalidate, this must be used in favour of cache
flushing so as not to degrade performance hence why the call must be
exposed to client drivers.  Now client drivers wishing to read back
buffer contents can now use invalidate instead of flush and if the
processor does not support an invalidate operation, we will issue a
flush (i.e. clean + invalidate) operation instead; for any perfected
data, the invalidate operation discards it from the processor cache.

In conclusion, client drivers should use CLEAN to transfer ownership
of the buffer to the device and use INVALIDATE to transfer ownership
back to processor.

Note that this change itself does not fix the bug below, but the
bugfixes herein are a prerequisite for later performance work.

BUG=chrome-os-partner:47837
TEST=Boot to UI, run WebGL aquarium.

Change-Id: I651d5f32890c7f6701f1191b4c376b474b9d1c22
Signed-off-by: Rufus Hamade <rufus.hamade@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/340523
Commit-Ready: Mao Huang <littlecvr@chromium.org>
Tested-by: Mao Huang <littlecvr@chromium.org>
Tested-by: Horng-Shyang Liao <hs.liao@mediatek.com>
Reviewed-by: Mao Huang <littlecvr@chromium.org>