CHROMIUM: drm/img-rogue: set parent clk to clk26m when changing freq
MT8173 GPU clk is top_mfg. top_mfg is the output of a mux, which can
have several inputs such as clk26m and mmpll.  Usually mmpll is
selected as top_mfg's parent to allow GPU DVFS to select a range of
frequencie for the GPU.

Currently, when changing the GPU frequency, we directly change mmpll
frequency.  However, it turns out that mmpll is unstable while its
frequency is changing.  This instability causes issues with the GPU
and/or for the CPU if happens to be attempting to access GPU registers.

To fix this, this patch first moves top_mfg over to a stable clk26m,
then changes the mmpll frequency, and finally then moves back to the
mmpll after mmpll is stable at the new frequency.

Signed-off-by: HS Liao <hs.liao@mediatek.com>

BUG=chrome-os-partner:52206
TEST=pass RunIn.Stress test

Change-Id: I61950b886bd9cc9fc556b4db97e71f5b60c419d6
Reviewed-on: https://chromium-review.googlesource.com/340662
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>