CHROMIUM: ath10k: Fix call traces during rx without proper channel context
WARN_ON_ONCE when we receive packets for Self peer when mac80211
had not assigned a proper channel context. This scenario happens
in QCA4019 when we start the AP via hostapd in background and start
it once again in the background without killing the previous instance!
This happens intermittently when we start / stop hostapd in a while loop
(incase the hostapd is not properly killed).  This results in mac80211
chancontext to be unassigned, while the self peer continuous receive
packets in target operating channel. This results in lot of call traces
in the rx path. Make this as a WARN_ON_ONCE to avoid flooding the console
which result in rebooting low memory systems, while still reporting the
warning once that we are receiving packets in target operating channel and
to indicate that something is happening which is not the expected result.

 WARNING: CPU: 0 PID: 0 at
ath/ath10k/htt_rx.c:803
[<c0318838>] (warn_slowpath_null) from [<bf4a0104>]
(ath10k_htt_rx_h_channel+0xe0/0x1b8 [ath10k_core])
[<bf4a0104>] (ath10k_htt_rx_h_channel [ath10k_core]) from
[<bf4a025c>] (ath10k_htt_rx_h_ppdu+0x80/0x288 [ath10k_core])
[<bf4a025c>] (ath10k_htt_rx_h_ppdu [ath10k_core]) from
[<bf4a1a9c>] (ath10k_htt_txrx_compl_task+0x724/0x9d4 [ath10k_core])
[<bf4a1a9c>] (ath10k_htt_txrx_compl_task [ath10k_core])
from [<c031be04>] (tasklet_action+0x8c/0xe8)
[<c031be04>] (tasklet_action) from [<c031b588>]
(__do_softirq+0xe0/0x204)

BUG=chrome-os-partner:49260
TEST=check if the warning shows up only once if hostapd is not killed properly

Change-Id: I471682de0159299ff3deba95e8876a9a015fc986
Signed-off-by: Mohammed Shafi Shajakhan <mohammed@codeaurora.org>
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/345733
Commit-Ready: Srinivasa duvvuri <sduvvuri@chromium.org>
Tested-by: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>