CHROMIUM: x86, apic: Add timed freeze support
This adds support to the clock event devices created by apic to use
timed freeze. The apic is able to run a timer during freeze with near
izero impact on modern CPUs such as skylake. This will allow S0ix,
suspend-to-idle, to be validated on Intel CPUs that support it.

This is needed because bugs with power settings on the SoC can prevent
S0ix entry. There is also no way to check this before idling all of
the CPUs.

BUG=chrome-os-partner:51441
TEST=powerd_dbus_suspend

Change-Id: I771d3aff8a4a1356d411c556da63d4cb7ee4805a
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/343277
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>