UPSTREAM: drm/i915/gen9: Allow calculation of data rate for in-flight state (v2)
Our skl_get_total_relative_data_rate() function gets passed a crtc state
object to calculate the data rate for, but it currently always looks
up the committed plane states that correspond to that CRTC.  Let's
check whether the CRTC state is an in-flight state (meaning
cstate->state is non-NULL) and if so, use the corresponding in-flight
plane states.

We'll soon be using this function exclusively for in-flight states; at
that time we'll be able to simplify the function a bit, but for now we
allow it to be used in either mode.

v2:
 - Rebase on top of changes to cache plane data rates.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1463061971-19638-5-git-send-email-matthew.d.roper@intel.com

BUG=chrome-os-partner:45364,chrome-os-partner:52008
TEST=Hot plugging 4K displays via HDMI/DP to Skl systems should not cause
TEST=flicker.

(cherry picked from commit 9c74d82621c553f0da1f41bd5d90f5eab5659fdb)
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Change-Id: I31402f268a231669b4ac17a488be80d10ad38429
Reviewed-on: https://chromium-review.googlesource.com/344383
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>