UPSTREAM: drm/i915: Track whether an atomic transaction changes the active CRTC's
For the purposes of DDB re-allocation we need to know whether a
transaction changes the list of CRTC's that are active.  While
state->modeset could be used for this purpose, that would be slightly
too aggressive since it would lead us to re-allocate the DDB when a
CRTC's mode changes, but not its final active state.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
Link: http://patchwork.freedesktop.org/patch/msgid/1463061971-19638-7-git-send-email-matthew.d.roper@intel.com

BUG=chrome-os-partner:45364,chrome-os-partner:52008
TEST=Hot plugging 4K displays via HDMI/DP to Skl systems should not cause
TEST=flicker.

(cherry picked from commit 8b4a7d0597cd9910d7127ffae6ae91d21853a8a2)
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>
Change-Id: I785ce645cc5a19714bc520e898c104d61cd8e034
Reviewed-on: https://chromium-review.googlesource.com/344385
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>