CHROMIUM: net: ipv4: Fix uninitialized flowi4_uid fields
In the course of debugging another problem, it was discovered that the UID
fields on temporary flows created for inbound packets by the routing code
were not being set.  The rule matching logic was therefore relying on
uninitialized data from the stack to make decisions.  Set these fields to
INVALID_UID so they are not checked.

IPv6 was lightly tested and did not appear to have the same issue.

BUG=b:28705881
TEST=manual test using an instrumented kernel

Change-Id: I9803e9b2d2e5fb067e964c56f0375e5fc8d068dd
Signed-off-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/344346
Reviewed-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit 773ec67588ae5977329bd4d59751528370cba4fc)
Reviewed-on: https://chromium-review.googlesource.com/362871