CHROMIUM: drm/img-rogue: Always print ERROR and FATAL messages
These are error messages, we ALWAYS want to see error messages, even
when building without the rest of the pvr debug infrastructure.

The simplest way to enable this is to catch PVR_DBG_FATAL and
PVR_DBG_ERROR messages in the PVR_DPF() macro, and print them with
pr_err().

To do this, we have to always define the PVR_DBG_* level macros.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:53750
TEST=On oak: Verify no extra non-error spam in kernel message log

Change-Id: I879e934e5e8dbe1d2c9bc0e3afbd0ddfb37c8682
Reviewed-on: https://chromium-review.googlesource.com/347540
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Horng-Shyang Liao <hs.liao@mediatek.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Horng-Shyang Liao <hs.liao@mediatek.com>