CHROMIUM: arm64: dts: mt8173-oak: Switch the function of the volume buttons
On mt8173-oak we set up the volume buttons in laptop mode perspective,
which causes the volume buttons to work the other way around when in
tablet mode. Since in laptop mode user has the volume keys on the
keyboard to control volume, the volume buttons on the side of the device
are mainly for tablet mode use and we should switch the function of the
two volume buttons.

BUG=chrome-os-partner:53755
TEST=Verify on my Elm that in tablet mode the volume buttons on the side
     of the device work as intended.

Change-Id: Ia609d7dc07f770bbe40bc1985e465a0fdd7ea273
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/347523
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>