CHROMIUM: ASoC: ipq4019: I2S startup sequence change
When the Gale device boots and loads the driver,
a pop noise is heard. The pop is also heard when the device
is rebooted (just before reset) or whenever any
audio application (e.g. aplay) is used for playback.

To fix this pop noise issue, the hardware initialisation
sequence has been modified as follows.
   1. GCC_AUDIO_BCR register is asserted by default.
   2. BCLK is turned off in depthcharge because it needs
      done before the TLMM initialisation.
   3. Enabling Stereo config.
   4. Enabling BCLK and MCLK.
   5. Starting the MBOX DMA.

BUG=chrome-os-partner:53767 chrome-os-partner:53768
TEST=unit tested audio driver with below test cases.
   1. Tested for static noise during the bootup.
   2. Tested for noise during the start up of the
      audio.

Change-Id: Ieba925608be12a6fb36ff67c7079ab9edfcaa2fa
Signed-off-by: Nitheesh Sekar <nsekar@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/348181
Commit-Ready: Suresh Rajashekara <sureshraj@chromium.org>
Tested-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>