CHROMIUM: arm64: add "model name" field in cpuinfo
On ARM platform uname in gentoo coreutils looks for a line starting
with "model name" in /proc/cpuinfo to determine the processor type.
Exporting "model name" in /proc/cpuinfo so that debugd can get the
cpu type for chrome://system.

This patch is very similar to

  https://patchwork.kernel.org/patch/9144985/

which enables "model name" in /proc/cpuinfo only in compact mode.
Since the cpuinfo related codes have changed a lot on upstream it's
not worth the effort to cherry-pick a lot of patches just for one
line of information in /proc/cpuinfo.

BUG=chrome-os-partner:53520
TEST=Run `uname -p` on my DUT.

Change-Id: I4dc3b31680906b1a7e17130f7db2620176b958ab
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348124
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>