CHROMIUM: tty: serial: msm: Check for ongoing serial DMA in console write
The kernel hangs sometimes if the console write happens with normal
TTY layer write and DMA is enabled. The current msm serial driver
uses the DMA, if the number of characters to be written is greater
than FIFO size. The console write uses the normal FIFO mode for all
its write operation. If console write is being called after
submitting the request to BAM DMA and before completion of this
request then the console write will not find TX FIFO empty in
following scenario

1. The kernel printk disables the interrupt before calling console
   write function by spin_lock_irqsave.

2. The DMA interrupt CPU affinity is same as the CPU in which
   console write function is running.

3. The DMA completion interrupt would not be triggered since it
   is disabled and TX FIFO empty condition will wait forever.

In this case, there is no other option except dropping the current
console write request to prevent kernel from hanging.

Now, this patch adds the check for DMA count and returns from console
write function if it is non zero.

BUG=chrome-os-partner:49255
TEST=Tested the simultaneous console and tty write

Change-Id: I62e15ffa2242b954079165bce71e013b62bda5b7
Signed-off-by: Abhishek Sahu <absahu@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/349451
Commit-Ready: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>