CHROMIUM: arm64: dts: mt8173-elm: Add ps8640_fixed_1v2 ramp delay
This fixed regulator actually takes 2 ms to ramp up when enabled.

If we don't wait for the regulator to settle and take the PS8640
out of reset too early, sometimes the bridge will fail to start
properly, and the eDP panel will stay black.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:53825
TEST=boot to ui on elm
TEST=Use magnet over F1 to repeatedly trigger lid close/open events
  Note: magnet must stay over magnet for ~0.5 to avoid triggering
   crbug.com/620208
  => display comes on every time

Change-Id: I7685ea804b49f7d22239957bcd2869088f0c98a4
Reviewed-on: https://chromium-review.googlesource.com/352491
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: cawa cheng <cawa.cheng@mediatek.com>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: jitao shi <jitao.shi@mediatek.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: jitao shi <jitao.shi@mediatek.com>