HACK: drm/nouveau: use trylock for cli->mutex during system suspend
The cli->mutex should not be held when the system enters system suspend.
To avoid potential locking issue which are caused by misused cli->mutex,
we use trylock in the system suspend patch. And if we do fail to acquire
the lock, the system suspend will be aborted.

BUG=chrome-os-partner:54432
TEST=Boot to UI

Change-Id: Ic271985e313784b6fe7fcb8916cd59953c89c24b
Signed-off-by: Vince Hsu <vince.h@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/353501
Trybot-Ready: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Stephen Barber <smbarber@chromium.org>
Tested-by: Stephen Barber <smbarber@chromium.org>