CHROMIUM: ASoC: ipq4019: dma_resume only when dma started
MBOX DMA resume is meaningful only in the following situations:
	1. DMA is started; MBOX engine has exhausted the queued descriptors;
	   and has gone into a STALLED state since there are no more
	   descriptors for it to process.
			(or)
	2. DMA is active

When DMA is resumed in state '1', MBOX resumes fetching and processing the
descriptors. And, in state '2', the resume is ignored.

Initiating a DMA resume when the MBOX is not in one of the above states can
potentially confuse the MBOX state machine causing it to refetch previous
descriptors and play them resulting in static/random noise. Hence, added
a flag to track if the MBOX is already started before resuming the DMA.

[This is a rework of https://chromium-review.googlesource.com/#/c/353515 .
 Kudos to Balamurugan Selvarajan <bselvara@codeaurora.org> for debugging
 and proposing the initial fix.]

BUG=chrome-os-partner:53767
TEST=unit tested audio driver with below test cases.
       1. Tested for static noise during playback
       2. Tested for noise during the start/end of the audio
       3. Loop audio playback for 500 times. Should play without
          random noise, crash or hangs.

Change-Id: Ib17052cf004b3f5a3eaad3553af06e4ea87abbd2
Signed-off-by: Grant Grundler <grundler@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/354084
Tested-by: Balamurugan Selvarajan <bselvara@codeaurora.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Balamurugan Selvarajan <bselvara@codeaurora.org>