CHROMIUM: arm64: dts: mt8173-oak: Change touchscreen interrupt to LEVEL_LOW
The Elan touchscreen interrupt is asserted (low) when the touchscreen has
data to send, and brought high when touch data has been read.  The
interrupt, therefore, is a low level interrupt, not a falling edge
interrupt.

Furthermore, this seems to work-around a bug in the MTK GPIO driver, which
seems to be triggering occasional double edge-triggered interrupts.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:53843
TEST=move finger around touchscreen and lift & replace often
 => No messages like this in /var/log/messages:
 elants_i2c 6-0010: unknown packet ff ff ff ff

Change-Id: I98b2de26033e4227365d6c5fb1f299541a9bb746
Reviewed-on: https://chromium-review.googlesource.com/354370
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Wei-Ning Huang <wnhuang@chromium.org>
Tested-by: Johnny Chuang <johnny.chuang@emc.com.tw>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
Reviewed-by: Johnny Chuang <johnny.chuang@emc.com.tw>