CHROMIUM: ath10k: hard code hw_rev enum value of QCA4019
Currently, wireless-4.2 stack in brillo doesn't have the latest ath10k
source from upstream. In upstream, there are additional hardwares/chipsets
supported by ath10k. With this addition, the hardware revision (hw_rev)
enum value for QCA4019 is moved to 6 from 4.

Hardware revision (hw_rev) helps to determine the firmware variant sending
these trace events.

It is important to have this hw_rev in sync with upstream thereby we can
use the upstream tracing plugin to decode trace data.

So fixing the hw_rev of QCA4019 to 6. Added a note to remove this when we
upgrade the wireless stack of 3.18

BUG=chrome-os-partner:53772
TEST=Build, bring up wifi and basic sanity test. Use uptramed tracing
plugin and verified decoding of pktlog works

Change-Id: I040d3d648a284b9878eb1b793c8ccbfa726dfe96
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/355372
Commit-Ready: Srinivasa duvvuri <sduvvuri@chromium.org>
Tested-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>