UPSTREAM: Input: add SW_PEN_INSERTED define
Some devices with a pen may have a switch that can be used to detect
when the pen is inserted or removed to a slot on the device.  Let's add
a define to the input event codes so that everyone can be on the same
page for what event we should generate when the pen is inserted or
removed.

In general the pen switch could be used by the software on the device to
kick off any number of actions when the pen is inserted or removed.

BUG=chrome-os-partner:54116
TEST=evtest

Change-Id: Ia942c41f0198925970fb5ae2a96e0d7ff80db5c3
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
(cherry picked from git.kernel.org/pub/scm/linux/kernel/git/dtor/input.git for-linus
 commit 9a9b6aa6a8759c83024627d681eff982d6ee03b7)
Reviewed-on: https://chromium-review.googlesource.com/356670
(cherry picked from commit 29a173512ed3bfe4e345cde5c91fa2adaefca642)
Conflicts:
	include/uapi/linux/input-event-codes.h
	since this file is missing on 3.18
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/425193
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>