FROMLIST: mtd: spi-nor: perform read-back check after SR lock/unlock
When programming the Status Register for performing write protect, it's
important to know the result of the operation (i.e., success or
failure). Particularly, it's possible to fail when the hardware write
protect pin (WP#) is asserted, potentially disallowing writes to the
status register. Previously, even if we weren't able to update the
status register, ioctl(MEMLOCK) and ioctl(MEMUNLOCK) would return
success.

Let's add a read-back check, to make sure that the status register was
updated appropriately. If the relevant bits weren't updated, then return
-EIO.

Signed-off-by: Brian Norris <computersforpeace@gmail.com>
(am from https://patchwork.kernel.org/patch/9228505/)

BUG=chrome-os-partner:54040, chrome-os-partner:54736
TEST=`flashrom --wp-disable` when WP screw is in

Change-Id: I6d7270e7266df3159f5e15f787879ba24fb314e3
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/358163
Reviewed-by: Julius Werner <jwerner@chromium.org>