UPSTREAM: Input: elan_i2c - fix calculating number of x and y traces.
According to Elan's firmware engineers we should not be subtracting 1 form
the raw number of x and y traces so that the pitch size is correct. For
example, if the touchpad x resolution is 2800 and x trace number is 20,
the pitch size of x should be 2800/20 = 140, not 2800/19 = 147.36.

Signed-off-by: Duson Lin <dusonlin@emc.com.tw>
Signed-off-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
(cherry picked from commit 9f423801c9e849df56b37833df895685c9d8a940)
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=none
TEST=build elm kernel with #define DEBUG in elan_i2c_core.
TEST=Confirm the "Width X,Y:" values in dmesg have changed,
     from "160,162" to "152,148".

Change-Id: Ie8551d3c8a4676e575a5e847be6b3e7036089f46
Reviewed-on: https://chromium-review.googlesource.com/360007
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>