CHROMIUM: drm/mediatek: remove target line interrupt
Previously, the command queue could only send one cmdq_rec per vblank.
This limited the drm driver to only issue one atomic update per vblank.
This is correct behavior for page flips, but it meant that the driver
could not perform a legacy cursor ioctl and page flip in the same vblank,
since that would require to atomic updates.  Instead, the display driver
collected commands from plane updates (primary and cursor) into
"cmdq_pending" variables.  The driver would then issue a single cmdq_rec
with a wait of vblank to apply all pending changes to both primary and
cursor planes.  To ensure that this cmdq_rec was issued before vblank, the
RDMA line interrupt was used to trigger an irq at 80% of frame duration.
This irq would then schedule the cmdq work to actually create and submit
the cmdq_rec to the command queue.

The major disadvantage of this scheme, is that using the line irq throws
away 20% of each frame.  That is, if a page flip is requested after the
line irq but before vblank, it would not get applied at the next vblank,
instead it would be queued up and must wait as long as 120% of a frame
period  before it appearing on screen.  Likewise, this would actually
reduce frame rate, as userspace would also have to wait before it
could submit the next page flip.

To fix this, the the command queue was updated to chain together atomic
commits and apply them en masse at vblank.  Therefore, the drm driver
need no longer need to use the RDMA line irq to carefully time and apply
cmdq_rec at 80% of frame duration.

This also allows us to implement atomic updates as they were intended -
plane updates can now be directly bundled in cmdq_recs and sent to the
cmdq without queuing up pending changes and applying them later in a
worker.

Changes to directly send cmdq commands in atomic_flush are implemented in
a preceding patch.  This patch cleans up the leftover RDMA and line_irq
interrupt handling code.

This patch depends on the patch:
"CHROMIUM: soc: mediatek: CMDQ: cascade tasks to wait one vsync for display"

Signed-off-by: Bibby Hsieh <bibby.hsieh@mediatek.com>
Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:54665
TEST=boot to ui on elm
TEST=HDMI work normal (plug/unplug/replug, display normal, show/hide cursor)
TEST=backlight on/off
TEST=suspend resume stress test finished 10000 iterations.
TEST=chameleon display_HotPlugAtSuspend.extended_unplugged
TEST=chameleon display_HotPlugAtSuspend.mirrored_unplugged

Change-Id: I9c63952109af9c6309d3c562db9922fa1c284e59
Reviewed-on: https://chromium-review.googlesource.com/354940
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Bibby Hsieh <bibby.hsieh@mediatek.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>