CHROMIUM: drm/mediatek: Only block updates to CRTCs that have a pending update
Currently we use a single mutex to allow only a single atomic update at a
time. In truth, though, we really only want to ensure that only a single
atomic update is allowed per CRTC.

In other words, for each atomic update, we only block if there is a
pending update for the CRTCs involved, and don't block if there are
only pending updates for other CRTCs.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:54665
TEST=boot to ui on elm
TEST=HDMI work normal (plug/unplug/replug, display normal, show/hide cursor)
TEST=backlight on/off
TEST=suspend resume stress test finished 10000 iterations.
TEST=chameleon display_HotPlugAtSuspend.extended_unplugged
TEST=chameleon display_HotPlugAtSuspend.mirrored_unplugged

Change-Id: I49c53921ccb08636533e5a684173bd1cabcf9301
Reviewed-on: https://chromium-review.googlesource.com/362293
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Bibby Hsieh <bibby.hsieh@mediatek.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>