CHROMIUM: ath10k: fix sending frame in management path in push txq logic
In the wake tx queue path, we are not checking if the frame to be sent
takes management path or not. For eg. QOS null func frame coming here will
take the management path. Since we are not incrementing the descriptor
counter (num_pending_mgmt_tx) w.r.t tx management, on tx completion it is
possible to see negative values.

When the above counter reaches a negative value, we will not be sending a
probe response out.

    if (is_presp &&
	ar->hw_params.max_probe_resp_desc_thres < htt->num_pending_mgmt_tx)

For IPQ4019, max_probe_resp_desc_thres (u32) is 24 is compared against
num_pending_mgmt_tx (int) and the above condtions comes true if the counter
is negative and we drop the probe response.

To avoid this, check on the wake tx queue path as well for the tx path of
the frame and increment the appropriate counters

BUG=chrome-os-partner:55146
TEST=bring up mesh and ap interface. See that the probe responses are not
dropped because of num_pending_mgmt_tx counter going negative

Change-Id: I0be2a61dc30ff96b2b9cfe1d7ff7584d18e78f65
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/365509
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Srinivasa duvvuri <sduvvuri@chromium.org>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>