CHROMIUM: drm/img-rogue: Change handle from MULTI to SHARED
Buffer sync expects user space to provide a set of unique PMRs to the
Services kernel driver but this was not happening as a PMR can
potentially have multiple handles, thus preventing user space from
filtering out all duplicates before passing them to the kernel.

This resulted in an EALREADY error being returned when trying to lock the
same PMR for a second time.

BUG=chrome-os-partner:56257
TEST=android.media.cts.EncodeVirtualDisplayTest

Change-Id: Ibadcbcdb5abeb1defb92444d5fde58a84e34744c
Signed-off-by: Jones Chen <jones.chen@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/374818
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>