CHROMIUM: PRIVATE: remove tx_status check for SA tx feedback processing
Currently, when the host receives a enhanced htt stats message, host checks
the tx_status from ppdu_stats in enhanced stats message. Host process only
if the tx status is zero i.e success.

Due to this check, when there is a complete failure of the packet due to
smart antenna training, SA algorithm doesn't know about this failure as we
are bailing out much earlier prior to tx feedback parsing. According to FW,
he has given a tx feedback with training bit set for failures in training.
So there is a mismatch in the packets counter in host and firmware.

This will cause training to not finish. 1. Due to mismatch in the counter
and the FW's counter will be exhausted that it won't send any more training
feedbacks while the host thinks it has not recived enought train packets to
progress in the antenna selection.

So remove this check.

BUG=None
TEST=Connect a station after enabling SA algorithm, and check if SA
training is happening in high noisy environment

Change-Id: I741d2239e208381db1a0d6263b86f6012fde4ea3
Signed-off-by: Ashok Raj Nagarajan <arnagara@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/386543
Commit-Ready: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>
Tested-by: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>
Reviewed-by: Oliver Wen <oliverwen@chromium.org>
Reviewed-by: Ashok Raj Nagarajan <Arnagara@qti.qualcomm.com>