CHROMIUM: Input: atmel_mxt_ts - put into deep sleep at shutdown on Celes
Celes does not cut power to the touch controller when system is shut
down, causing issues when we try to reinitialize the controller when
system is booted up again.

To work around this issue we put the touch controller into deep sleep at
shutdown. Ideally firmware would do that for as (because kernel's
expectations is that peripherals are powered off when we shut the system
down, but at this time kernel workaround is easiest).

BUG=chrome-os-partner:57704
TEST=build kernel.
     check if atmel ic goes to deep sleep mode at power off.
     Power On, check if touch screen/touchpad work at power on.

Change-Id: I45de8f981482d560860d448b2c7f440fad477e19
Signed-off-by: Jongpil Jung <jongpil19.jung@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/390211
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>