CHROMIUM: drm/mediatek: Synchronize access to cmdq_vblank_event and cmdq_needs_event
It is possible on an SMP system for CPU0 executing the ddp_cmdq_cb isr to
read the new value for cmdq_needs_event (1) and an old value for
cmdq_vblank_event (NULL) leading to an OOPs in
drm_crtc_send_vblank_event().

Plug this situation using a pair of memory barriers.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:57705
TEST=sanity check UI
TEST=24 hours:
  https://www.khronos.org/registry/webgl/sdk/demos/webkit/Earth.html

Change-Id: Icb8aa9da42a43bd2f8ea846c76820f0ea566a283
Reviewed-on: https://chromium-review.googlesource.com/392589
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>