CHROMIUM: drm/img-rogue: Rename the 1.6 driver
Add a '_1_6' suffix to the 1.6 driver name in preparation for upgrading to
a new driver release. This will allow for both old and new kernel drivers to
co-exist while transitioning to a new usermode driver. This is necessary as
currently the kernel/user interface isn't stable across different releases.

BUG=chrome-os-partner:58144
TEST=Boot to UI

Change-Id: Ibe962870a61c527fec0fef811ddef135be439fae
Signed-off-by: Frank Binns <frank.binns@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/411821
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>