CHROMIUM: bluetooth: adjust advertising duration based on max interval
The advertising duration determines how long an advertisement is
advertised. As the default duration is 2000 ms, the controller will
keep advertising the same advertisement continuously in the duration
until the duration times out and next advertisement instance is
scheduled. This behavior makes shorter advertising intervals useless.

This patch adjusts the duration according to the max advertising interval
so that multiple advertisements could be interleaved in a round-robin
manner. Besides, a millisecond is used as the unit of duration instead of
a second for finer resolution.

BUG=chromium:655251
TEST=Run the chrome advertising test app, use nRF Connect, and watch that
the advertisement instances are interleaved in a round-robin manner.

Change-Id: Id436029c6a11116d09c1f01ea2bfe0e0047de015
Reviewed-on: https://chromium-review.googlesource.com/400202
Commit-Ready: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Wei-Ning Huang <wnhuang@chromium.org>
(cherry picked from commit 768362c86a0e84d52cde438effa1d7721b8f300a)
Reviewed-on: https://chromium-review.googlesource.com/401305
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>