CHROMIUM: config: arm64: panic on hardlockup detect
We've had this set to panic-on-hardlockup for a long time for arm32.
Let's do that for arm64 too.

This helps for debugging and collecting crash information, since
otherwise a hardlockup is not guaranteed to trigger the panic handlers.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>
Signed-off-by: Brian Norris <briannorris@chromium.org>

From chromeos-4.4: https://chromium-review.googlesource.com/396396

BUG=chromium:654931
TEST=On elm:
  `echo HARDLOCKUP > /sys/kernel/debug/provoke-crash/DIRECT`
 => see panic instead of indefinite hang
TEST=platform_KernelErrorPaths.HARDLOCKUP
  => test passes

Change-Id: If994b18b40b546aeeb3ca3da20f911416a46fe1c
Reviewed-on: https://chromium-review.googlesource.com/400125
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>